Update according to removing 10sec timer
