+++ test_scripts/Handling_VR_help_requests/001_SetGlobalProp_by_adding_commands.lua
@@ -8,14 +8,9 @@
 -- Description:
 -- In case:
 -- 1. Command1, Command2, Command3 commands with vrCommands are added
--- 2. 10 seconds timer is expired
--- 3. SDL sends SetGlobalProperties  with constructed the vrHelp and helpPrompt parameters using added vrCommand.
--- 4. Mobile application sets 30 command one by one
--- 5.Mobile application deletes 10 command one by one
--- 6. Mobile application sets 31 command
 -- SDL does:
--- send SetGlobalProperties  with update for vrHelp and helpPrompt parameters after each added and deleted command
--- not send SetGlobalProperties after added 31 command
+-- send SetGlobalProperties  with constructed the vrHelp and helpPrompt parameters using added vrCommands
+-- after receiving each command
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
@@ -33,17 +28,9 @@ runner.Step("App activation", common.activateApp)
 
 runner.Title("Test")
 for i = 1,3 do
-  runner.Step("AddCommand" .. i, common.addCommand, { common.getAddCommandParams(i) })
+  runner.Step("SetGlobalProperties after AddCommand" .. i, common.addCommandWithSetGP, { i })
 end
-runner.Step("SetGlobalProperties with constructed the vrHelp and helpPrompt", common.setGlobalPropertiesFromSDL,
-  { true })
-for i = 4, 33 do
-  runner.Step("SetGlobalProperties from SDL after added command" ..i, common.addCommandWithSetGP, { i })
-end
-for i = 1, 10 do
-  runner.Step("SetGlobalProperties from SDL after deleted command" ..i, common.deleteCommandWithSetGP, { i })
-end
-runner.Step("Absence SetGlobalProperties from SDL after adding 34 command", common.addCommandWithoutSetGP, { 34 })
+runner.Step("Absence of SetGlobalProperties request from SDL", common.setGlobalPropertiesDoesNotExpect)
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
