+++ test_scripts/Handling_VR_help_requests/004_SetGlobalProp_processing_30_commands.lua
@@ -7,11 +7,11 @@
 --
 -- Description:
 -- In case:
--- 1. Command1 commands with vrCommands Command1_1, Command2_1, Command3_1 is added
--- 2. Perform reconnect
--- 2. 10 seconds timer is expired
+-- 1. Mobile application sets 30 command one by one
+-- 2. Mobile application sets 31 command
 -- SDL does:
--- send SetGlobalProperties with constructed the vrHelp and helpPrompt parameters using added vrCommands.
+-- 1. send SetGlobalProperties with full list of command values for vrHelp and helpPrompt parameters after each added command
+-- 2. not send SetGlobalProperties after added 31 command
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
@@ -20,25 +20,18 @@ local common = require('test_scripts/Handling_VR_help_requests/commonVRhelp')
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
 
---[[ Local Variables ]]
-local AddCommandParams = common.getAddCommandParams(1)
-AddCommandParams.vrCommands = { "Command1_1", "Command2_1", "Command3_1" }
-
 --[[ Scenario ]]
 runner.Title("Preconditions")
 runner.Step("Clean environment", common.preconditions)
 runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
 runner.Step("App registration", common.registerAppWOPTU)
-runner.Step("Pin OnHashChange", common.pinOnHashChange)
 runner.Step("App activation", common.activateApp)
 
 runner.Title("Test")
-runner.Step("AddCommand with several vrCommand", common.addCommand, { AddCommandParams })
-runner.Step("App reconnect", common.reconnect)
-runner.Step("App resumption", common.registrationWithResumption,
-  { 1, common.resumptionLevelFull, common.resumptionDataAddCommands })
-runner.Step("SetGlobalProperties with constructed the vrHelp and helpPrompt", common.setGlobalPropertiesFromSDL,
-	{ true })
+for i = 1, 30 do
+  runner.Step("SetGlobalProperties from SDL after added command" ..i, common.addCommandWithSetGP, { i })
+end
+runner.Step("Absence SetGlobalProperties from SDL after adding 31 command", common.addCommandWithoutSetGP, { 31 })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
