+++ test_scripts/Handling_VR_help_requests/006_SetGlobalProp_processing_added_choices.lua
@@ -8,10 +8,10 @@
 -- Description:
 -- In case:
 -- 1. Command1, Command2, Command3 commands with vrCommands are added
--- 2. Mobile application sets SetGlobalProperties with custom vrHelp
--- 3. Mobile application adds  Command4
+-- 2. vrCommands Choice1, Choice2 are added via CreateInterationChoiceSet
 -- SDL does:
--- 1. send SetGlobalProperties  with updated value for the helpPrompt parameter using added vrCommands
+-- send SetGlobalProperties with constructed the vrHelp and helpPrompt parameters using added vrCommands via AddCommand
+-- requests(with type "Command" ).
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
@@ -20,21 +20,37 @@ local common = require('test_scripts/Handling_VR_help_requests/commonVRhelp')
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
 
---[[ Local Variables ]]
-local SetGPParamsWithVrHelpOnly = common.customSetGPParams()
-SetGPParamsWithVrHelpOnly.requestParams.helpPrompt = nil
+-- [[ Local Variables ]]
+local requestParams = {
+  interactionChoiceSetID = 1001,
+  choiceSet = {
+    {
+      choiceID = 1001,
+      menuName ="Choice1001",
+      vrCommands = {
+	    "Choice1001_1", "Choice1001_2"
+      }
+    }
+  }
+}
 
---[[ Local Functions ]]
-local function SetGlobalPropertiesFromSDLbyAddingCommand()
-  common.addCommand(common.getAddCommandParams(4))
-  local params = common.getGPParams()
+local responseVrParams = {
+  cmdID = requestParams.interactionChoiceSetID,
+  type = "Choice",
+  vrCommands = requestParams.vrCommands
+}
+
+-- [[ Local Functions ]]
+local function createInteractionChoiceSetWithoutSetGP()
+  local mobSession = common.getMobileSession(1)
   local hmiConnection = common.getHMIConnection()
-  EXPECT_HMICALL("UI.SetGlobalProperties")
-  :Times(0)
-  EXPECT_HMICALL("TTS.SetGlobalProperties", params.requestTtsParams)
+  local cid = mobSession:SendRPC("CreateInteractionChoiceSet", requestParams)
+  EXPECT_HMICALL("VR.AddCommand", responseVrParams)
   :Do(function(_,data)
     hmiConnection:SendResponse(data.id, data.method, "SUCCESS", {})
   end)
+  mobSession:ExpectResponse(cid, { success = true, resultCode = "SUCCESS"})
+  common.setGlobalPropertiesDoesNotExpect()
 end
 
 --[[ Scenario ]]
@@ -43,15 +59,12 @@ runner.Step("Clean environment", common.preconditions)
 runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
 runner.Step("App registration", common.registerAppWOPTU)
 runner.Step("App activation", common.activateApp)
+
+runner.Title("Test")
 for i = 1,3 do
   runner.Step("AddCommand" .. i, common.addCommandWithSetGP, { i })
 end
-
-runner.Title("Test")
-runner.Step("Custom SetGlobalProperties with vrHelp only from mobile application", common.setGlobalProperties,
-  { SetGPParamsWithVrHelpOnly })
-runner.Step("SetGlobalProperties with updated value for helpPrompt after added command ",
-  SetGlobalPropertiesFromSDLbyAddingCommand)
+runner.Step("CreateInteractionChoiceSet", createInteractionChoiceSetWithoutSetGP)
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
