+++ test_scripts/Handling_VR_help_requests/008_SetGlobalProp_absence_in_BACKGROUND.lua
@@ -6,9 +6,11 @@
 -- Requirement summary: TBD
 --
 -- Description:
+-- Description:
 -- In case:
--- 1. App is registered and set in NONE HMI level.
--- 2. PT is updated with "NONE" for AddCommand RPC
+-- 1. App is registered and activated.
+-- 2. PT is updated with "BACKGROUND" for AddCommand RPC
+-- 3. App deactivated to BACKGROUND HMI level
 -- 3. Command1, Command2, Command3 commands with vrCommands are added
 -- SDL does:
 -- send SetGlobalProperties  with constructed the vrHelp and helpPrompt parameters using added vrCommands
@@ -28,12 +30,27 @@ local function ptuFunc(tbl)
 	tbl.policy_table.functional_groupings["Base-4"].rpcs.AddCommand.hmi_levels = { "BACKGROUND", "FULL", "LIMITED", "NONE" }
 end
 
+local function activateApp()
+  local requestId = common.getHMIConnection():SendRequest("SDL.ActivateApp", { appID = common.getHMIAppId(1) })
+  common.getHMIConnection():ExpectResponse(requestId)
+  common.getMobileSession(1):ExpectNotification("OnHMIStatus",
+    { hmiLevel = "FULL", audioStreamingState = "NOT_AUDIBLE", systemContext = "MAIN" })
+end
+
+local function deactivateApp()
+  common.getHMIConnection():SendNotification("BasicCommunication.OnAppDeactivated", {appID = common.getHMIAppId(1)})
+  common.getMobileSession(1):ExpectNotification("OnHMIStatus",
+    {hmiLevel = "BACKGROUND", audioStreamingState = "NOT_AUDIBLE", systemContext = "MAIN"})
+end
+
 --[[ Scenario ]]
 runner.Title("Preconditions")
 runner.Step("Clean environment", common.preconditions)
 runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
 runner.Step("App1 registration", common.registerApp)
-runner.Step("PTU with NONE hmi level for AddCommand", common.policyTableUpdate, { ptuFunc })
+runner.Step("App activation", activateApp)
+runner.Step("PTU with BACKGROUND hmi level for AddCommand", common.policyTableUpdate, { ptuFunc })
+runner.Step("App deactivation", deactivateApp)
 
 runner.Title("Test")
 for i = 1,3 do
