+++ test_scripts/Handling_VR_help_requests/009_SetGlobalProp_absence_addCommand_after_custom_request.lua
@@ -8,12 +8,9 @@
 -- Description:
 -- In case:
 -- 1. Command1, Command2, Command3 commands with vrCommands are added
--- 2. Mobile application sets SetGlobalProperties with custom helpPrompt and vrHelp
--- 3. 10 seconds timer is expired and SDL send SetGlobalProperties  with constructed the vrHelp and helpPrompt parameters using added vrCommands
--- 4. Mobile application deletes Command1 command
--- 5.SDL sends SetGlobalProperties with updated list of command
--- 6.Mobile application sets SetGlobalProperties with custom helpPrompt and vrHelp
--- 7.Mobile application deletes Command2
+-- 2. SDL sends SetGlobalProperties with updated list of command
+-- 3. Mobile application sets SetGlobalProperties with custom helpPrompt and vrHelp
+-- 4. Mobile application adds Command4 command
 -- SDL does:
 -- not send SetGlobalProperties with updated values for the vrHelp and helpPrompt parameters
 ---------------------------------------------------------------------------------------------------
@@ -31,18 +28,14 @@ runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
 runner.Step("App registration", common.registerAppWOPTU)
 runner.Step("App activation", common.activateApp)
 for i = 1,3 do
-  runner.Step("AddCommand" .. i, common.addCommand, { common.getAddCommandParams(i) })
+  runner.Step("AddCommand" .. i, common.addCommandWithSetGP, { i })
 end
-runner.Step("SetGlobalProperties with constructed the vrHelp and helpPrompt", common.setGlobalPropertiesFromSDL,
-  { true })
 
 runner.Title("Test")
-runner.Step("SetGlobalProperties from SDL with updated values after deleting command1", common.deleteCommandWithSetGP,
-  { 1 })
 runner.Step("Custom SetGlobalProperties from mobile application", common.setGlobalProperties,
   { common.customSetGPParams() })
-runner.Step("Absence of SetGlobalProperties request from SDL after deleting command2", common.deleteCommandWithoutSetGP,
-  { 2 })
+runner.Step("Absence of SetGlobalProperties request from SDL after added command4", common.addCommandWithoutSetGP,
+  { 4 })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
