+++ test_scripts/Handling_VR_help_requests/010_SetGlobalProp_absence_deleteCommand_after_custom_request.lua
@@ -7,11 +7,12 @@
 --
 -- Description:
 -- In case:
--- 1. Command1, Command2, Command3 commands with vrCommands are added
--- 2. Command1 is deleted by app
--- 3. 10 seconds timer is expired
+-- 1. Mobile application deletes Command1 command
+-- 2. SDL sends SetGlobalProperties with updated list of command
+-- 3. Mobile application sets SetGlobalProperties with custom helpPrompt and vrHelp
+-- 4. Mobile application deletes  Command2
 -- SDL does:
--- send SetGlobalProperties with constructed the vrHelp and helpPrompt parameters using added vrCommand.
+-- not send SetGlobalProperties with updated values for the vrHelp and helpPrompt parameters
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
@@ -20,14 +21,6 @@ local common = require('test_scripts/Handling_VR_help_requests/commonVRhelp')
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
 
---[[ Local Functions ]]
-local function getDeleteCommandParams()
-  local out = {
-    cmdID = 1
-  }
-  return out
-end
-
 --[[ Scenario ]]
 runner.Title("Preconditions")
 runner.Step("Clean environment", common.preconditions)
@@ -35,13 +28,16 @@ runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
 runner.Step("App registration", common.registerAppWOPTU)
 runner.Step("App activation", common.activateApp)
 for i = 1,3 do
-  runner.Step("AddCommand" .. i, common.addCommand, { common.getAddCommandParams(i) })
+  runner.Step("AddCommand" .. i, common.addCommandWithSetGP, { i })
 end
 
 runner.Title("Test")
-runner.Step("Delete Command1", common.deleteCommand, { getDeleteCommandParams(), true })
-runner.Step("SetGlobalProperties with constructed the vrHelp and helpPrompt", common.setGlobalPropertiesFromSDL,
-  { true })
+runner.Step("SetGlobalProperties from SDL with updated values after deleting command1", common.deleteCommandWithSetGP,
+  { 1 })
+runner.Step("Custom SetGlobalProperties from mobile application", common.setGlobalProperties,
+  { common.customSetGPParams() })
+runner.Step("Absence of SetGlobalProperties request from SDL after deleting command2", common.deleteCommandWithoutSetGP,
+  { 2 })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
