+++ test_scripts/Handling_VR_help_requests/017_SetGlobalProp_without_deleted_commands_by_resumption.lua
@@ -8,11 +8,12 @@
 -- Description:
 -- In case:
 -- 1. Command1, Command2, Command3 commands with vrCommands are added
--- 2. Perform reconnect
--- 3. Mobile application sets SetGlobalProperties without helpPrompt and vrHelp
--- 4. 10 seconds timer is expired
+-- 2. Command1 is deleted by app
+-- 3. Perform reopening session
 -- SDL does:
--- send SetGlobalProperties with constructed the vrHelp and helpPrompt parameters using added vrCommand with type=Command.
+-- 1. resume HMI level and AddCommands
+-- 2. send SetGlobalProperties with constructed the vrHelp and helpPrompt parameters using added vrCommand
+--  after each resumed command
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
@@ -20,16 +21,6 @@ local common = require('test_scripts/Handling_VR_help_requests/commonVRhelp')
 
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
-
---[[ Local Variables ]]
-local setGPParams = { }
-setGPParams.requestParams = {
-  keyboardProperties = {
-	keyboardLayout = "QWERTY",
-	keypressMode = "SINGLE_KEYPRESS"
-  }
-}
-setGPParams.requestUiParams = setGPParams.requestParams
 
 --[[ Scenario ]]
 runner.Title("Preconditions")
@@ -39,17 +30,14 @@ runner.Step("App registration", common.registerAppWOPTU)
 runner.Step("Pin OnHashChange", common.pinOnHashChange)
 runner.Step("App activation", common.activateApp)
 for i = 1,3 do
-  runner.Step("AddCommand" .. i, common.addCommand, { common.getAddCommandParams(i) })
+  runner.Step("AddCommand" .. i, common.addCommandWithSetGP, { i })
 end
+
+runner.Title("Test")
+runner.Step("Delete command1", common.deleteCommandWithSetGP, { 1 })
 runner.Step("App reconnect", common.reconnect)
 runner.Step("App resumption", common.registrationWithResumption,
   { 1, common.resumptionLevelFull, common.resumptionDataAddCommands })
-
-runner.Title("Test")
-runner.Step("Custom SetGlobalProperties from mobile application without helpPrompt and vrHelp",
-	common.setGlobalProperties, { setGPParams })
-runner.Step("SetGlobalProperties request from SDL with constructed the vrHelp and helpPrompt",
-	common.setGlobalPropertiesFromSDL, { true })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
