+++ test_scripts/Handling_VR_help_requests/023_SetGlobalProp_custom_vrHelp_only_by_resumption.lua
@@ -8,13 +8,11 @@
 -- Description:
 -- In case:
 -- 1. Command1, Command2, Command3 commands with vrCommands are added
--- 2. Perform reconnect
--- 3. Mobile application sets SetGlobalProperties with custom helpPrompt
--- 4. 10 seconds timer is expired
--- 5. Mobile application adds  Command4
+-- 3. Perform reconnect
+-- 4. Mobile application sets SetGlobalProperties with custom vrHelp
+-- 5. Mobile application adds Command4
 -- SDL does:
--- 1. send SetGlobalProperties  with constructed the vrHelp parameter using added vrCommand.
--- 2. send SetGlobalProperties  with updated value for the vrHelp parameter using added vrCommand
+-- 1. send SetGlobalProperties with updated value for the helpPrompt parameter using added vrCommands
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
@@ -24,32 +22,20 @@ local common = require('test_scripts/Handling_VR_help_requests/commonVRhelp')
 runner.testSettings.isSelfIncluded = false
 
 --[[ Local Variables ]]
-local SetGPParamsWithHelpPromptOnly = common.customSetGPParams()
-SetGPParamsWithHelpPromptOnly.requestParams.vrHelpTitle = nil
-SetGPParamsWithHelpPromptOnly.requestParams.vrHelp = nil
+local SetGPParamsWithVrHelpOnly = common.customSetGPParams()
+SetGPParamsWithVrHelpOnly.requestParams.helpPrompt = nil
 
 --[[ Local Functions ]]
-local function SetGlobalPropertiesFromSDL()
-  local params = common.getGPParams()
-  local hmiConnection = common.getHMIConnection()
-  EXPECT_HMICALL("UI.SetGlobalProperties", params.requestUiParams)
-  :Do(function(_,data)
-    hmiConnection:SendResponse(data.id, data.method, "SUCCESS", {})
-  end)
-  EXPECT_HMICALL("TTS.SetGlobalProperties")
-  :Times(0)
-end
-
 local function SetGlobalPropertiesFromSDLbyAddingCommand()
   common.addCommand(common.getAddCommandParams(4))
   local params = common.getGPParams()
   local hmiConnection = common.getHMIConnection()
-  EXPECT_HMICALL("UI.SetGlobalProperties", params.requestUiParams)
+  EXPECT_HMICALL("UI.SetGlobalProperties")
+  :Times(0)
+  EXPECT_HMICALL("TTS.SetGlobalProperties", params.requestTtsParams)
   :Do(function(_,data)
     hmiConnection:SendResponse(data.id, data.method, "SUCCESS", {})
   end)
-  EXPECT_HMICALL("TTS.SetGlobalProperties")
-  :Times(0)
 end
 
 --[[ Scenario ]]
@@ -67,10 +53,9 @@ runner.Step("App resumption", common.registrationWithResumption,
   { 1, common.resumptionLevelFull, common.resumptionDataAddCommands })
 
 runner.Title("Test")
-runner.Step("Custom SetGlobalProperties with helpPrompt only from mobile application", common.setGlobalProperties,
-  { SetGPParamsWithHelpPromptOnly })
-runner.Step("SetGlobalProperties with constructed the vrHelp", SetGlobalPropertiesFromSDL)
-runner.Step("SetGlobalProperties with updated value for vrHelp after added command ",
+runner.Step("Custom SetGlobalProperties with vrHelp only from mobile application", common.setGlobalProperties,
+  { SetGPParamsWithVrHelpOnly })
+runner.Step("SetGlobalProperties with updated value for helpPrompt after added command",
   SetGlobalPropertiesFromSDLbyAddingCommand)
 
 runner.Title("Postconditions")
