+++ test_scripts/Handling_VR_help_requests/026_SetGlobalProp_with_several_vrCommands_in_one_addCommand_request_by_resumption.lua
@@ -7,10 +7,12 @@
 --
 -- Description:
 -- In case:
--- 1. Command1, Command2, Command3 commands with vrCommands are added
--- 2. 10 seconds timer is expired
+-- 1. Command1 commands with vrCommands Command1_1, Command2_1, Command3_1 is added
+-- 2. Perform reconnect
+-- 3. SDL resumes commands
 -- SDL does:
--- send SetGlobalProperties with constructed the vrHelp and helpPrompt parameters using added vrCommands.
+-- send SetGlobalProperties with constructed the vrHelp and helpPrompt parameters using added vrCommands
+-- by command resumption
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
@@ -19,19 +21,23 @@ local common = require('test_scripts/Handling_VR_help_requests/commonVRhelp')
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
 
+--[[ Local Variables ]]
+local AddCommandParams = common.getAddCommandParams(1)
+AddCommandParams.vrCommands = { "Command1_1", "Command2_1", "Command3_1" }
+
 --[[ Scenario ]]
 runner.Title("Preconditions")
 runner.Step("Clean environment", common.preconditions)
 runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
 runner.Step("App registration", common.registerAppWOPTU)
+runner.Step("Pin OnHashChange", common.pinOnHashChange)
 runner.Step("App activation", common.activateApp)
 
 runner.Title("Test")
-for i = 1,3 do
-  runner.Step("AddCommand" .. i, common.addCommand, { common.getAddCommandParams(i) })
-end
-runner.Step("SetGlobalProperties with constructed the vrHelp and helpPrompt", common.setGlobalPropertiesFromSDL,
-  { true })
+runner.Step("AddCommand with several vrCommand", common.addCommandWithSetGP, { nil, AddCommandParams })
+runner.Step("App reconnect", common.reconnect)
+runner.Step("App resumption", common.registrationWithResumption,
+  { 1, common.resumptionLevelFull, common.resumptionDataAddCommands })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
