+++ test_scripts/Handling_VR_help_requests/028_SetGlobalProp_rejected_custom_request.lua
@@ -8,8 +8,9 @@
 -- Description:
 -- In case:
 -- 1. Command1, Command2, Command3 commands with vrCommands are added
--- 2. Mobile app deletes Command3 and HMI responds with resultCode = REJECTED, as result command is not deleted
--- 3. 10 seconds timer is expired
+-- 2. Mobile application sends valid SetGlobalProperties and HMI rejected request
+-- 3. SDL responds with resultCode REJECTED to mobile application
+-- 4. Mobile application adds Command4
 -- SDL does:
 -- send SetGlobalProperties  with constructed the vrHelp and helpPrompt parameters using added vrCommands.
 ---------------------------------------------------------------------------------------------------
@@ -20,28 +21,19 @@ local common = require('test_scripts/Handling_VR_help_requests/commonVRhelp')
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
 
---[[ Local Variables ]]
-local params = {
-    cmdID = 3
-  }
-
 --[[ Local Functions ]]
-local function rejectedDeleteCommand(pParams)
+local function rejectedSetGlobalProperties()
   local mobSession = common.getMobileSession()
   local hmiConnection = common.getHMIConnection()
-  local cid = mobSession:SendRPC("DeleteCommand", pParams)
-  EXPECT_HMICALL("UI.DeleteCommand")
+  local Params = common.customSetGPParams()
+  local cid = mobSession:SendRPC("SetGlobalProperties", Params.requestParams)
+  EXPECT_HMICALL("UI.SetGlobalProperties")
   :Do(function(_,data)
-    hmiConnection:SendError(data.id, data.method, "REJECTED", "Request rejected")
+    hmiConnection:SendError(data.id, data.method, "REJECTED", " UI is rejected ")
   end)
-  local requestVrParams = {
-    cmdID = pParams.cmdID,
-    type = "Command",
-    appID = common.getHMIAppId()
-  }
-  EXPECT_HMICALL("VR.DeleteCommand", requestVrParams)
+  EXPECT_HMICALL("TTS.SetGlobalProperties")
   :Do(function(_,data)
-    hmiConnection:SendError(data.id, data.method, "REJECTED", "Request rejected")
+    hmiConnection:SendError(data.id, data.method, "REJECTED", " TTS is rejected ")
   end)
   mobSession:ExpectResponse(cid, { success = false, resultCode = "REJECTED"})
 end
@@ -53,13 +45,12 @@ runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
 runner.Step("App registration", common.registerAppWOPTU)
 runner.Step("App activation", common.activateApp)
 for i = 1,3 do
-  runner.Step("AddCommand" .. i, common.addCommand, { common.getAddCommandParams(i) })
+  runner.Step("AddCommand" .. i, common.addCommandWithSetGP, { i })
 end
 
 runner.Title("Test")
-runner.Step("Rejected deleting Command3", rejectedDeleteCommand, { params })
-runner.Step("SetGlobalProperties with constructed the vrHelp and helpPrompt", common.setGlobalPropertiesFromSDL,
-	{ true })
+runner.Step("Rejected custom SetGlobalProperties", rejectedSetGlobalProperties)
+runner.Step("SetGlobalProperties after AddCommand 4", common.addCommandWithSetGP, { 4 })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
