+++ test_scripts/Handling_VR_help_requests/029_SetGlobalProp_only_successful_added_commands.lua
@@ -8,11 +8,9 @@
 -- Description:
 -- In case:
 -- 1. Command1, Command2, Command3 commands with vrCommands are added
--- 2. Mobile application sends valid SetGlobalProperties and HMI rejected request
--- 3. SDL responds with resultCode REJECTED to mobile application
--- 4. 10 seconds timer is expired
+-- 2. Mobile app adds Command4 and HMI responds with resultCode = REJECTED, as result command is not added
 -- SDL does:
--- send SetGlobalProperties  with constructed the vrHelp and helpPrompt parameters using added vrCommands.
+-- not send SetGlobalProperties  with constructed the vrHelp and helpPrompt parameters using added vrCommands by processing Command4.
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
@@ -22,18 +20,31 @@ local common = require('test_scripts/Handling_VR_help_requests/commonVRhelp')
 runner.testSettings.isSelfIncluded = false
 
 --[[ Local Functions ]]
-local function rejectedSetGlobalProperties()
+local function rejectedAddCommand(pParams)
   local mobSession = common.getMobileSession()
   local hmiConnection = common.getHMIConnection()
-  local Params = common.customSetGPParams()
-  local cid = mobSession:SendRPC("SetGlobalProperties", Params.requestParams)
+  local cid = mobSession:SendRPC("AddCommand", pParams)
+
   EXPECT_HMICALL("UI.SetGlobalProperties")
-  :Do(function(_,data)
-    hmiConnection:SendError(data.id, data.method, "REJECTED", " UI is rejected ")
-  end)
+  :Times(0)
+
   EXPECT_HMICALL("TTS.SetGlobalProperties")
+  :Times(0)
+
+  EXPECT_HMICALL("UI.AddCommand")
   :Do(function(_,data)
-    hmiConnection:SendError(data.id, data.method, "REJECTED", " TTS is rejected ")
+    hmiConnection:SendError(data.id, data.method, "REJECTED", "Rejected request")
+  end)
+  local requestUiParams = {
+    cmdID = pParams.cmdID,
+    vrCommands = pParams.vrCommands,
+    type = "Command",
+    appID = common.getHMIAppId()
+  }
+
+  EXPECT_HMICALL("VR.AddCommand", requestUiParams)
+  :Do(function(_,data)
+    hmiConnection:SendError(data.id, data.method, "REJECTED", "Rejected request")
   end)
   mobSession:ExpectResponse(cid, { success = false, resultCode = "REJECTED"})
 end
@@ -45,13 +56,11 @@ runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
 runner.Step("App registration", common.registerAppWOPTU)
 runner.Step("App activation", common.activateApp)
 for i = 1,3 do
-  runner.Step("AddCommand" .. i, common.addCommand, { common.getAddCommandParams(i) })
+  runner.Step("AddCommand" .. i, common.addCommandWithSetGP, { i })
 end
 
 runner.Title("Test")
-runner.Step("Rejected custom SetGlobalProperties", rejectedSetGlobalProperties)
-runner.Step("SetGlobalProperties with constructed the vrHelp and helpPrompt", common.setGlobalPropertiesFromSDL,
-	{ true })
+runner.Step("Rejected adding Command4", rejectedAddCommand, { common.getAddCommandParams(4) })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
