+++ test_scripts/Handling_VR_help_requests/030_SetGlobalProp_only_without_successful_deleted_commands.lua
@@ -8,10 +8,9 @@
 -- Description:
 -- In case:
 -- 1. Command1, Command2, Command3 commands with vrCommands are added
--- 2. Mobile app adds Command4 and HMI responds with resultCode = REJECTED, as result command is not added
--- 3. 10 seconds timer is expired
+-- 2. Mobile app deletes Command3 and HMI responds with resultCode = REJECTED, as result command is not deleted
 -- SDL does:
--- send SetGlobalProperties  with constructed the vrHelp and helpPrompt parameters using added vrCommands.
+-- not send SetGlobalProperties  with constructed the vrHelp and helpPrompt parameters using added vrCommands by processing Command4.
 ---------------------------------------------------------------------------------------------------
 --[[ Required Shared libraries ]]
 local runner = require('user_modules/script_runner')
@@ -20,24 +19,36 @@ local common = require('test_scripts/Handling_VR_help_requests/commonVRhelp')
 --[[ Test Configuration ]]
 runner.testSettings.isSelfIncluded = false
 
+--[[ Local Variables ]]
+local params = {
+    cmdID = 3
+  }
+
 --[[ Local Functions ]]
-local function rejectedAddCommand(pParams)
+local function rejectedDeleteCommand(pParams)
   local mobSession = common.getMobileSession()
   local hmiConnection = common.getHMIConnection()
-  local cid = mobSession:SendRPC("AddCommand", pParams)
-  EXPECT_HMICALL("UI.AddCommand")
+  local cid = mobSession:SendRPC("DeleteCommand", pParams)
+
+  EXPECT_HMICALL("UI.SetGlobalProperties")
+  :Times(0)
+
+  EXPECT_HMICALL("TTS.SetGlobalProperties")
+  :Times(0)
+
+  EXPECT_HMICALL("UI.DeleteCommand")
   :Do(function(_,data)
-    hmiConnection:SendError(data.id, data.method, "REJECTED", "Rejected request")
+    hmiConnection:SendError(data.id, data.method, "REJECTED", "Request rejected")
   end)
-  local requestUiParams = {
+  local requestVrParams = {
     cmdID = pParams.cmdID,
-    vrCommands = pParams.vrCommands,
     type = "Command",
     appID = common.getHMIAppId()
   }
-  EXPECT_HMICALL("VR.AddCommand", requestUiParams)
+
+  EXPECT_HMICALL("VR.DeleteCommand", requestVrParams)
   :Do(function(_,data)
-    hmiConnection:SendError(data.id, data.method, "REJECTED", "Rejected request")
+    hmiConnection:SendError(data.id, data.method, "REJECTED", "Request rejected")
   end)
   mobSession:ExpectResponse(cid, { success = false, resultCode = "REJECTED"})
 end
@@ -49,13 +60,11 @@ runner.Step("Start SDL, HMI, connect Mobile, start Session", common.start)
 runner.Step("App registration", common.registerAppWOPTU)
 runner.Step("App activation", common.activateApp)
 for i = 1,3 do
-  runner.Step("AddCommand" .. i, common.addCommand, { common.getAddCommandParams(i) })
+  runner.Step("AddCommand" .. i, common.addCommandWithSetGP, { i })
 end
 
 runner.Title("Test")
-runner.Step("Rejected adding Command4", rejectedAddCommand, { common.getAddCommandParams(4) })
-runner.Step("SetGlobalProperties with constructed the vrHelp and helpPrompt", common.setGlobalPropertiesFromSDL,
-	{ true })
+runner.Step("Rejected deleting Command3", rejectedDeleteCommand, { params })
 
 runner.Title("Postconditions")
 runner.Step("Stop SDL", common.postconditions)
