+++ test_sets/handling_VR_help_requests.txt
@@ -1,31 +1,30 @@
-./test_scripts/Handling_VR_help_requests/001_SetGlobalProp_with_added_commands_after_timeout.lua
+./test_scripts/Handling_VR_help_requests/001_SetGlobalProp_by_adding_commands.lua
 ./test_scripts/Handling_VR_help_requests/002_SetGlobalProp_commands_with_and_without_vrCommand.lua
-./test_scripts/Handling_VR_help_requests/003_SetGlobalProp_without_deleted_commands_after_timeout.lua
-./test_scripts/Handling_VR_help_requests/004_SetGlobalProp_processing_30_commands_after_timeout.lua
-./test_scripts/Handling_VR_help_requests/005_SetGlobalProp_apply_limitation_after_timeout_only_for_addcommand.lua
-./test_scripts/Handling_VR_help_requests/006_SetGlobalProp_procesing_added_choices.lua
+./test_scripts/Handling_VR_help_requests/003_SetGlobalProp_without_deleted_commands.lua
+./test_scripts/Handling_VR_help_requests/004_SetGlobalProp_processing_30_commands.lua
+./test_scripts/Handling_VR_help_requests/005_SetGlobalProp_apply_limitation_only_for_addcommand.lua
+./test_scripts/Handling_VR_help_requests/006_SetGlobalProp_processing_added_choices.lua
 ./test_scripts/Handling_VR_help_requests/007_SetGlobalProp_absence_in_NONE.lua
-./test_scripts/Handling_VR_help_requests/008_SetGlobalProp_absence_after_custom_request.lua
-./test_scripts/Handling_VR_help_requests/009_SetGlobalProp_absence_addCommand_after_custom_request_after_timeout.lua
-./test_scripts/Handling_VR_help_requests/010_SetGlobalProp_absence_deleteCommand_after_custom_request_after_timeout.lua
+./test_scripts/Handling_VR_help_requests/008_SetGlobalProp_absence_in_BACKGROUND.lua
+./test_scripts/Handling_VR_help_requests/009_SetGlobalProp_absence_addCommand_after_custom_request.lua
+./test_scripts/Handling_VR_help_requests/010_SetGlobalProp_absence_deleteCommand_after_custom_request.lua
 ./test_scripts/Handling_VR_help_requests/011_SetGlobalProp_custom_helpPrompt_only.lua
 ./test_scripts/Handling_VR_help_requests/012_SetGlobalProp_custom_vrHelp_only.lua
 ./test_scripts/Handling_VR_help_requests/013_SetGlobalProp_custom_helpPrompt_and_vrHelp_in_different_requests.lua
 ./test_scripts/Handling_VR_help_requests/014_SetGlobalProp_custom_request_without_helpPrompt_and_vrHelp.lua
-./test_scripts/Handling_VR_help_requests/015_SetGlobalProp_with_added_commands_after_timeout_by_resumption.lua
+./test_scripts/Handling_VR_help_requests/015_SetGlobalProp_with_added_commands_by_resumption.lua
 ./test_scripts/Handling_VR_help_requests/016_SetGlobalProp_commands_with_and_without_vrCommand_by_resumption.lua
-./test_scripts/Handling_VR_help_requests/017_SetGlobalProp_without_deleted_commands_after_timeout_by_resumption.lua
-./test_scripts/Handling_VR_help_requests/018_SetGlobalProp_processing_30_commands_after_timeout_by_resumption.lua
-./test_scripts/Handling_VR_help_requests/019_SetGlobalProp_procesing_added_choices_by_resumption.lua
+./test_scripts/Handling_VR_help_requests/017_SetGlobalProp_without_deleted_commands_by_resumption.lua
+./test_scripts/Handling_VR_help_requests/018_SetGlobalProp_processing_30_commands_by_resumption.lua
+./test_scripts/Handling_VR_help_requests/019_SetGlobalProp_processing_added_choices_by_resumption.lua
 ./test_scripts/Handling_VR_help_requests/020_SetGlobalProp_absence_after_custom_request_after_resumption.lua
 ./test_scripts/Handling_VR_help_requests/021_SetGlobalProp_absence_after_custom_request_before_resumption.lua
-./test_scripts/Handling_VR_help_requests/022_SetGlobalProp_absence_after_constructed_vrHelp_helpPrompt_before_resumption.lua
-./test_scripts/Handling_VR_help_requests/023_SetGlobalProp_custom_helpPrompt_only_by_resumption.lua
-./test_scripts/Handling_VR_help_requests/024_SetGlobalProp_custom_vrHelp_only_by_resumption.lua
-./test_scripts/Handling_VR_help_requests/025_SetGlobalProp_custom_request_without_helpPrompt_and_vrHelp_by_resumption.lua
-./test_scripts/Handling_VR_help_requests/026_SetGlobalProp_with_several_vrCommands_in_one_addCommand_request.lua
-./test_scripts/Handling_VR_help_requests/027_SetGlobalProp_with_several_vrCommands_in_one_addCommand_request_by_resumption.lua
-./test_scripts/Handling_VR_help_requests/028_SetGlobalProp_invalid_custom_request.lua
-./test_scripts/Handling_VR_help_requests/029_SetGlobalProp_rejected_custom_request.lua
-./test_scripts/Handling_VR_help_requests/030_SetGlobalProp_only_successful_added_commands.lua
-./test_scripts/Handling_VR_help_requests/031_SetGlobalProp_only_without_successful_deleted_commands.lua
+./test_scripts/Handling_VR_help_requests/022_SetGlobalProp_custom_helpPrompt_only_by_resumption.lua
+./test_scripts/Handling_VR_help_requests/023_SetGlobalProp_custom_vrHelp_only_by_resumption.lua
+./test_scripts/Handling_VR_help_requests/024_SetGlobalProp_custom_request_without_helpPrompt_and_vrHelp_by_resumption.lua
+./test_scripts/Handling_VR_help_requests/025_SetGlobalProp_with_several_vrCommands_in_one_addCommand_request.lua
+./test_scripts/Handling_VR_help_requests/026_SetGlobalProp_with_several_vrCommands_in_one_addCommand_request_by_resumption.lua
+./test_scripts/Handling_VR_help_requests/027_SetGlobalProp_invalid_custom_request.lua
+./test_scripts/Handling_VR_help_requests/028_SetGlobalProp_rejected_custom_request.lua
+./test_scripts/Handling_VR_help_requests/029_SetGlobalProp_only_successful_added_commands.lua
+./test_scripts/Handling_VR_help_requests/030_SetGlobalProp_only_without_successful_deleted_commands.lua
